refactor: Memo 네임스페이스 구현, attributes생성 기능 프로토콜 기본구현으로 수정
